date adapter: add hack to parse ISO8601 on Android
Currently, Android's implementation of the `SimpleDAteFormat` class does not
include the format char `X`. On desktop clients, the wildcard `X` token was
being used to ignore the GMT zone signifer `Z`.

This works fine on the desktop, since GMT just means +00:00. On Android,
however, an exception is thrown.

As a temporary workaround, the un-parsable 'Z' is replaced with "+0000", which
is parsable with the `Z` delimiter on both Android and Windows.